Skeleton Checkout

+++ internal/product/model/product.go
@@ -13,7 +13,7 @@ type Product struct {
 	Category           string       `json:"category" db:"category"`
 	ImageURL           string       `json:"imageUrl" db:"image_url"`
 	Notes              string       `json:"notes" db:"notes"`
-	Price              int64        `json:"price" db:"price"`
+	Price              float64      `json:"price" db:"price"`
 	Stock              int          `json:"stock" db:"stock"`
 	Location           string       `json:"location" db:"location"`
 	IsAvailable        bool         `json:"isAvailable" db:"is_available"`
